+++ src/index.ts
@@ -1,3 +1,4 @@
 import "./tokens/index.css";
 
 export { version } from "./version.ts";
+export * from "./components/index.ts";